requirements.txt
beautifulsoup4==4.11.1
coloredlogs==15.0.1
pandas==1.5.2
pydantic==1.10.2
pytest==7.1.2
python-dateutil==2.8.2
retry==0.9.2
requests=2.28.1